🎉 Release v1.0.0.0 - Bugs Fixed

+++ src/main/java/com/konsol/beatstream/service/audioPlugins/youtube/YoutubeDownloader.java
@@ -11,6 +11,7 @@ import com.konsol.beatstream.service.TrackService;
 import com.konsol.beatstream.service.UserService;
 import com.konsol.beatstream.service.audioPlugins.ReferanceTrack.ReferanceTrackHandler;
 import com.konsol.beatstream.service.bucket.BucketManager;
+import com.konsol.beatstream.service.impl.PlaylistServiceImpl;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
@@ -22,11 +23,15 @@ import org.jaudiotagger.audio.AudioFileIO;
 import org.jaudiotagger.audio.AudioHeader;
 import org.jaudiotagger.tag.FieldKey;
 import org.jaudiotagger.tag.Tag;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class YoutubeDownloader {
+
+    private static final Logger LOG = LoggerFactory.getLogger(YoutubeDownloader.class);
 
     @Autowired
     TrackService trackService;
@@ -52,7 +57,7 @@ public class YoutubeDownloader {
     private final String YTDLP_PATH = startupPath + "\\plugins\\yt-dlp.exe";
     private final String FFMPEG_PATH = startupPath + "\\plugins\\ffmpeg.exe";
 
-    public void AddYoutubeVideo(String video_Id, String playlistId) {
+    public void AddYoutubeVideo(String video_Id, String playlistId) throws Exception {
         this.videoId = video_Id;
 
         User user = userService.getCurrentUser();
@@ -63,34 +68,44 @@ public class YoutubeDownloader {
             manager.createBucket(user.getId());
             manager.createBucket(user.getId() + "\\" + "audioFiles");
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            LOG.debug(e.getMessage());
+            throw new Exception(e);
         }
         Path youtubeDownloadDirPath = rootPath.resolve("youTubeDL");
         String videoUrl = "https://www.youtube.com/watch?v=" + videoId; // Replace with actual URL
         String outputPath = youtubeDownloadDirPath + "\\" + videoId + "/" + videoId; // Output location and file format
 
         if (!YouTubeVideoInfoScraper.isValidYouTubeUrl(videoUrl)) {
-            throw new RuntimeException("Invalid YouTube Video URL: " + videoUrl);
+            LOG.debug("Invalid YouTube Video URL: " + videoUrl);
+            throw new Exception("Invalid YouTube Video URL: " + videoUrl);
+        }
+
+        if (videoId == null) {
+            LOG.debug("YouTube Video Id Is: null");
+            throw new Exception("YouTube Video Id Is: null");
         }
 
         if (trackRepository.findByRefIdAndRefTypeAndOwnerId(videoId, "YOUTUBE", user.getId()).isPresent()) {
-            throw new RuntimeException("YouTube Video already exists: " + videoId);
-        }
+            LOG.debug("YouTube Video already exists: " + videoId);
+            throw new Exception("YouTube Video already exists: " + videoId);
+        } else {
+            Track track = trackService.createTrack(videoId, "YOUTUBE", playlistId);
 
-        Track track = trackService.createTrack(videoId, "YOUTUBE", playlistId);
+            referanceDownloadTask = referanceTrackHandler.createDownloadTask(video_Id, "YOUTUBE", track.getId());
 
-        referanceDownloadTask = referanceTrackHandler.createDownloadTask(video_Id, "YOUTUBE", track.getId());
-
-        try {
-            downloadYouTubeVideoAsMP3(videoUrl, outputPath, track);
-        } catch (IOException | InterruptedException e) {
-            trackService.delete(track.getId());
+            try {
+                downloadYouTubeVideoAsMP3(videoUrl, outputPath, track);
+            } catch (Exception e) {
+                trackService.delete(track.getId());
+                throw e;
+            }
         }
     }
 
     public void downloadYouTubeVideoAsMP3(String videoUrl, String outputPath, Track track) throws IOException, InterruptedException {
         // Prepare the yt-dlp command
-        String command = YTDLP_PATH + " -x --audio-format mp3 --ffmpeg-location " + FFMPEG_PATH + " -o " + outputPath + " " + videoUrl;
+        String command =
+            "\"" + YTDLP_PATH + "\" -x --audio-format mp3 --ffmpeg-location \"" + FFMPEG_PATH + "\" -o " + outputPath + " " + videoUrl;
 
         // Run the command
         ProcessBuilder processBuilder = new ProcessBuilder("cmd", "/c", command);
@@ -105,7 +120,7 @@ public class YoutubeDownloader {
             referanceDownloadTask.setReferanceStatus("Downloading");
             referanceDownloadTask = referanceDownloadTaskRepository.save(referanceDownloadTask);
             while ((line = reader.readLine()) != null) {
-                System.out.println(line);
+                LOG.debug(line);
                 referanceDownloadTask.setReferanceLog(referanceDownloadTask.getReferanceLog() + "\n" + line);
                 referanceDownloadTask = referanceDownloadTaskRepository.save(referanceDownloadTask);
             }
@@ -129,13 +144,14 @@ public class YoutubeDownloader {
                 referanceDownloadTask.setReferanceStatus("AddingAudioToPlayList");
                 referanceDownloadTask = referanceDownloadTaskRepository.save(referanceDownloadTask);
             } catch (Exception e) {
+                LOG.debug("YouTube Video already exists: " + videoId);
                 trackService.delete(track.getId());
             }
             referanceDownloadTask.setReferanceStatus("DownloadComplete");
             referanceDownloadTask = referanceDownloadTaskRepository.save(referanceDownloadTask);
-            System.out.println("Download complete!");
+            LOG.debug("Download complete!");
         } else {
-            System.err.println("Download failed. Exit code: " + exitCode);
+            LOG.debug("Download failed. Exit code: " + exitCode);
         }
     }
 
@@ -143,14 +159,15 @@ public class YoutubeDownloader {
         try {
             trackService.connectTrackToAudioFile(track.getId(), outputPath.toString());
         } catch (Exception e) {
+            LOG.debug(e.getMessage());
             throw e;
         }
     }
 
-    public Track extrackAndSaveTrackData(Track track, Path outputPath, String videoUrl) {
+    public Track extrackAndSaveTrackData(Track track, Path outputPath, String videoUrl) throws Exception {
         if (!outputPath.toFile().exists()) {
             trackService.delete(track.getId());
-            throw new RuntimeException("File Not Found");
+            throw new Exception("File Not Found");
         }
         try {
             // Read the audio file and its metadata using jaudiotagger
@@ -177,7 +194,7 @@ public class YoutubeDownloader {
 
             return trackRepository.save(track);
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            throw new Exception(e);
         }
     }
 }
